@@ -6,7 +6,39 @@
  */
 
 // any CSS you import will output into a single css file (app.css in this case)
-import './styles/app.css';
+import './styles/app.scss';
 
 // start the Stimulus application
-import './bootstrap';
+import 'bootstrap';
+import 'bootstrap';
+
+import 'bootstrap-tagsinput';
+
+import Bloodhound from "bloodhound-js";
+import 'typeahead.js';
+
+const $ = require('jquery');
+
+$(function() {
+    // Bootstrap-tagsinput initialization
+    // https://bootstrap-tagsinput.github.io/bootstrap-tagsinput/examples/
+    var $input = $('input[data-toggle="tagsinput"]');
+    if ($input.length) {
+        var source = new Bloodhound({
+            local: $input.data('tags'),
+            queryTokenizer: Bloodhound.tokenizers.whitespace,
+            datumTokenizer: Bloodhound.tokenizers.whitespace
+        });
+        source.initialize();
+
+        $input.tagsinput({
+            trimValue: true,
+            focusClass: 'focus',
+            typeaheadjs: {
+                name: 'tags',
+                source: source.ttAdapter()
+            },
+            delimiter: ','
+        });
+    }
+});
